ErrorCode와 ErrorDetail의 value() 및 category() 변환형 변경
* ErrorCode와 ErrorDetail의 value() 및 category() 반환형을 uintptr_t에서
uint32_t로 변경

+++ include/bicomc/detail/error_code.h
@@ -72,12 +72,12 @@ namespace bcc
 		return mpDetail == nullptr;
 	}
 	
-	inline bcc::uintptr_t ErrorCode::value() const
+	inline bcc::uint32_t ErrorCode::value() const
 	{
 		return mpDetail ? mpDetail->value() : 0;
 	}
 
-	inline bcc::uintptr_t ErrorCode::category() const
+	inline bcc::uint32_t ErrorCode::category() const
 	{
 		return mpDetail ? mpDetail->category() : 0;
 	}
