+++ include/bicomc/detail/error_fwd.h
@@ -11,8 +11,8 @@ namespace bcc
 namespace detail
 {
 	BICOMC_INTERFACE(ErrorDetail)
-		BICOMC_DECL_METHOD_0_C(value, bcc::uintptr_t());
-		BICOMC_DECL_METHOD_0_C(category, bcc::uintptr_t());
+		BICOMC_DECL_METHOD_0_C(value, bcc::uint32_t());
+		BICOMC_DECL_METHOD_0_C(category, bcc::uint32_t());
 		BICOMC_DECL_METHOD_0_C(message, char const*());
 	BICOMC_INTERFACE_END(ErrorDetail);
 } // namespace detail
